# apps/cloud-api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TelaFlow Cloud API",
    "description": "API da TelaFlow Cloud \u2014 fase inicial: governan\u00e7a e integra\u00e7\u00e3o, sem persist\u00eancia completa.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/events/{event_id}/scenes": {
      "get": {
        "tags": [
          "scenes"
        ],
        "summary": "List Scenes",
        "description": "Lista scenes do evento, ordenadas por sort_order.",
        "operationId": "list_scenes_events__event_id__scenes_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Scenes Events  Event Id  Scenes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "scenes"
        ],
        "summary": "Create Scene",
        "description": "Cria scene; `scene_id` gerado no servidor; `sort_order` \u00fanico por evento.",
        "operationId": "create_scene_events__event_id__scenes_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Scene Events  Event Id  Scenes Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}/scenes/reorder": {
      "post": {
        "tags": [
          "scenes"
        ],
        "summary": "Reorder Scenes",
        "description": "Define a ordem das scenes: `scene_ids` deve listar todos os ids atuais do evento,\ncada um exatamente uma vez, na ordem desejada.\nAtribui `sort_order` de 0 a n-1 na ordem recebida.",
        "operationId": "reorder_scenes_events__event_id__scenes_reorder_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneReorderBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reorder Scenes Events  Event Id  Scenes Reorder Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}/scenes/{scene_id}": {
      "get": {
        "tags": [
          "scenes"
        ],
        "summary": "Get Scene",
        "operationId": "get_scene_events__event_id__scenes__scene_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Scene Events  Event Id  Scenes  Scene Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "scenes"
        ],
        "summary": "Update Scene",
        "operationId": "update_scene_events__event_id__scenes__scene_id__patch",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scene Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Scene Events  Event Id  Scenes  Scene Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "scenes"
        ],
        "summary": "Delete Scene",
        "operationId": "delete_scene_events__event_id__scenes__scene_id__delete",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Scene Events  Event Id  Scenes  Scene Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SceneCreate": {
        "properties": {
          "sort_order": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sort Order"
          },
          "type": {
            "type": "string",
            "enum": [
              "opening",
              "institutional",
              "sponsor",
              "draw",
              "break",
              "closing"
            ],
            "title": "Type"
          },
          "name": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "media_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Id"
          },
          "draw_config_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draw Config Id"
          }
        },
        "type": "object",
        "required": [
          "sort_order",
          "type",
          "name"
        ],
        "title": "SceneCreate",
        "description": "Cria\u00e7\u00e3o \u2014 `scene_id` e `event_id` definidos pela rota."
      },
      "SceneReorderBody": {
        "properties": {
          "scene_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scene Ids"
          }
        },
        "type": "object",
        "title": "SceneReorderBody",
        "description": "Nova ordem: lista completa de `scene_id` na sequ\u00eancia desejada (0 = primeiro)."
      },
      "SceneUpdate": {
        "properties": {
          "sort_order": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          },
          "type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "opening",
                  "institutional",
                  "sponsor",
                  "draw",
                  "break",
                  "closing"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "media_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Id"
          },
          "draw_config_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draw Config Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SceneUpdate",
        "description": "Atualiza\u00e7\u00e3o parcial (PATCH)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}